Fix cyclic import issue.
