+++ df_py/util/test/test_blocktime_eth_mainnet.py
@@ -8,7 +8,7 @@ from df_py.util.blocktime import (
     eth_timestamp_to_block,
     timestr_to_block,
 )
-from df_py.util.oceanutil import get_rpc_url, get_web3
+from df_py.util.web3 import get_rpc_url, get_web3
 
 
 @enforce_types
